requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --annotation-style=line requirements.in
#
annotated-types==0.6.0    # via pydantic
anthropic==0.21.3         # via -r requirements.in
anyio==4.3.0              # via anthropic, httpx
appium-python-client==4.0.0  # via -r requirements.in
attrs==23.2.0             # via outcome, trio
certifi==2024.2.2         # via httpcore, httpx, requests, selenium
charset-normalizer==3.3.2  # via requests
distro==1.9.0             # via anthropic
exceptiongroup==1.2.0     # via anyio, trio, trio-websocket
filelock==3.13.1          # via huggingface-hub
fsspec==2024.3.1          # via huggingface-hub
h11==0.14.0               # via httpcore, wsproto
httpcore==1.0.4           # via httpx
httpx==0.27.0             # via anthropic
huggingface-hub==0.21.4   # via tokenizers
idna==3.6                 # via anyio, httpx, requests, trio
outcome==1.3.0.post0      # via trio
packaging==24.0           # via huggingface-hub
pydantic==2.6.4           # via anthropic
pydantic-core==2.16.3     # via pydantic
pysocks==1.7.1            # via urllib3
pyyaml==6.0.1             # via huggingface-hub
requests==2.31.0          # via huggingface-hub
selenium==4.18.1          # via appium-python-client
sniffio==1.3.1            # via anthropic, anyio, httpx, trio
sortedcontainers==2.4.0   # via trio
tokenizers==0.15.2        # via anthropic
tqdm==4.66.2              # via huggingface-hub
trio==0.25.0              # via selenium, trio-websocket
trio-websocket==0.11.1    # via selenium
typing-extensions==4.10.0  # via anthropic, anyio, huggingface-hub, pydantic, pydantic-core, selenium
urllib3[socks]==2.2.1     # via requests, selenium
wsproto==1.2.0            # via trio-websocket
